+++ src/Commands/LogzSingle.php
@@ -31,8 +31,7 @@ class LogzSingle extends Command
    */
   //
   public function __construct(LogzHandler $logzHandler)
-  {
-    var_dump('2');
+  { 
     parent::__construct();
     $this->handler = $logzHandler;
   }